Basic search bar with component
This component is class based, a controlled component. This means that the component
gets its info from the state of the component, instead of the component dictating the state.
ie. the value of the input is determined by the state of the component, instead of the value
of the input dictating the state of the component.