Converting App to Class and Setting Initial State
We converted the App to a class based component and setup a constructor.
This constructor initializes the state and makes a YT request with a search term
and then updates this.state.videos with the list of videos. This way we can have
a list of videos for the user from the start.